added middleware authentication

+++ routes/web.php
@@ -20,9 +20,9 @@ use function PHPUnit\Framework\assertClassHasStaticAttribute;
 
 Route::get('/', [HomeController::class, 'index']);
 
-Route::get('/tasks', [TaskController::class, 'index']);
+Route::get('/tasks', [TaskController::class, 'index'])->middleware('is_admin');
 
-Route::get('tasks/create', [TaskController::class, 'create']);
+Route::get('tasks/create', [TaskController::class, 'create'])->middleware('is_admin');
 
 // Route::get('/task', function () use ($taskList) {
 //     if (request()->search) {
@@ -40,20 +40,20 @@ Route::get('/tasks/{id}', [TaskController::class, 'show']); //show
 //     return $taskList;
 // });
 
-Route::post('/tasks', [TaskController::class, 'store']); //store
+Route::post('/tasks', [TaskController::class, 'store'])->middleware('is_admin'); //store
 
 // Route::patch('/task/{key}', function ($key) use ($taskList) {
 //     $taskList[request()->key] = request()->task;
 //     return $taskList;
 // });
 
-Route::get('/tasks/{id}/edit', [TaskController::class, 'edit']);
+Route::get('/tasks/{id}/edit', [TaskController::class, 'edit'])->middleware('is_admin');
 
-Route::patch('/tasks/{id}', [TaskController::class, 'update']); //update
+Route::patch('/tasks/{id}', [TaskController::class, 'update'])->middleware('is_admin'); //update
 
 // Route::delete('/task/{key}', function ($key) use ($taskList) {
 //     unset($taskList[$key]);
 //     return $taskList;
 // });
 
-Route::delete('/tasks/{id}', [TaskController::class, 'delete']); //delete
+Route::delete('/tasks/{id}', [TaskController::class, 'delete'])->middleware('is_admin'); //delete
